+++ src/Umbraco.Web.UI.Client/src/packages/core/server-file-system/rename/index.ts
@@ -2,3 +2,4 @@ export type * from './types.js';
 export * from './rename-server-file.action.js';
 export * from './rename-server-file-repository-base.js';
 export * from './event/index.js';
+export * from './workspace-redirect/index.js';
