+++ src/Umbraco.Web.UI.Client/src/packages/core/server-file-system/rename/workspace-redirect/index.ts
@@ -1,0 +1,1 @@
+export * from './workspace-redirect.controller.js';
