+++ src/Umbraco.Web.UI.Client/src/packages/templating/scripts/workspace/script-workspace.context.ts
@@ -11,9 +11,7 @@ import {
 	UmbWorkspaceIsNewRedirectController,
 } from '@umbraco-cms/backoffice/workspace';
 import type { IRoutingInfo, PageComponent } from '@umbraco-cms/backoffice/router';
-import { UMB_ACTION_EVENT_CONTEXT } from '@umbraco-cms/backoffice/action';
-import { UmbServerFileRenamedEntityEvent } from '@umbraco-cms/backoffice/server-file-system';
-import { ensurePathEndsWithSlash, umbUrlPatternToString } from '@umbraco-cms/backoffice/utils';
+import { UmbServerFileRenameWorkspaceRedirectController } from '@umbraco-cms/backoffice/server-file-system';
 
 export class UmbScriptWorkspaceContext
 	extends UmbEntityDetailWorkspaceContextBase<UmbScriptDetailModel, UmbScriptDetailRepository>
@@ -22,22 +20,11 @@ export class UmbScriptWorkspaceContext
 	public readonly name = this._data.createObservablePartOfCurrent((data) => data?.name);
 	public readonly content = this._data.createObservablePartOfCurrent((data) => data?.content);
 
-	#actionEventContext?: typeof UMB_ACTION_EVENT_CONTEXT.TYPE;
-
 	constructor(host: UmbControllerHost) {
 		super(host, {
 			workspaceAlias: UMB_SCRIPT_WORKSPACE_ALIAS,
 			entityType: UMB_SCRIPT_ENTITY_TYPE,
 			detailRepositoryAlias: UMB_SCRIPT_DETAIL_REPOSITORY_ALIAS,
-		});
-
-		this.consumeContext(UMB_ACTION_EVENT_CONTEXT, (context) => {
-			this.#actionEventContext = context;
-
-			if (this.#actionEventContext) {
-				this.#actionEventContext.removeEventListener(UmbServerFileRenamedEntityEvent.TYPE, this.#onFileRenamed);
-				this.#actionEventContext.addEventListener(UmbServerFileRenamedEntityEvent.TYPE, this.#onFileRenamed);
-			}
 		});
 
 		this.routes.setRoutes([
@@ -62,6 +49,12 @@ export class UmbScriptWorkspaceContext
 				setup: (component: PageComponent, info: IRoutingInfo) => {
 					const unique = info.match.params.unique;
 					this.load(unique);
+
+					new UmbServerFileRenameWorkspaceRedirectController(
+						this,
+						this,
+						this.getHostElement().shadowRoot!.querySelector('umb-router-slot')!,
+					);
 				},
 			},
 		]);
@@ -83,32 +76,6 @@ export class UmbScriptWorkspaceContext
 	 */
 	public setContent(value: string) {
 		this._data.updateCurrent({ content: value });
-	}
-
-	#onFileRenamed = ((event: UmbServerFileRenamedEntityEvent) => {
-		const currentUnique = this.getUnique();
-		const eventUnique = event.getUnique();
-		if (currentUnique !== eventUnique) return;
-
-		const newUnique = event.getNewUnique();
-		if (!newUnique) throw new Error('New unique is required for this event.');
-
-		const router = this.getHostElement().shadowRoot!.querySelector('umb-router-slot')!;
-
-		if (router) {
-			const routerPath = router.absoluteRouterPath;
-			if (routerPath) {
-				const newPath: string = umbUrlPatternToString(ensurePathEndsWithSlash(routerPath) + 'edit/:unique', {
-					unique: newUnique,
-				});
-				window.history.pushState({}, '', newPath);
-			}
-		}
-	}) as EventListener;
-
-	public override destroy(): void {
-		super.destroy();
-		this.#actionEventContext?.removeEventListener(UmbServerFileRenamedEntityEvent.TYPE, this.#onFileRenamed);
 	}
 }
